requirements.txt
numpy~=1.20.1
pillow~=8.2.0
torch==1.10.2
torchvision==0.11.1
opencv-python~=4.5.2.52
matplotlib~=3.3.4
pandas~=1.2.4
seaborn~=0.11.1
scikit-learn~=0.24.1
flask~=1.1.2
werkzeug~=1.0.1
scikit-image~=0.18.1
h5py~=3.2.1
requests~=2.25.1
selenium~=3.141.0
flask_sqlalchemy~=2.5.1
imutils~=0.5.4
